intel-compute-runtime: Use libva-initial as build dependency
libva pulls in mesa as an implicit dependency, which may be
built with a different llvm version than intel-graphics-compiler,
and pulling in two different llvm SDK conflicts.

Use libva-initial to resolve this build time conflict.

There won't be a runtime conflict because the libLLVM library
files are versioned in their name.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>